[refactor] split files for classes
H5FileInfo and ChLoadAddedMass
