+++ CMakeLists.txt
@@ -58,14 +58,37 @@ endif()
 # Add the executable from your project and specify all C++ 
 # files in your project. 
 #--------------------------------------------------------------
+set(HYDRO_CHRONO_SOURCES
+  
+	src/h5fileinfo.cpp
+	src/chloadaddedmass.cpp
+	src/hydro_forces.cpp
 
-add_library(HydroChrono STATIC "src/hydro_forces.cpp" "src/hydro_forces.h")
-add_executable(sphere_decay "demos/sphere/sphere_decay.cpp")
-add_executable(sphere_reg_waves "demos/sphere/sphere_reg_waves.cpp")
-add_executable(rm3_decay "demos/rm3/rm3_decay.cpp")
-add_executable(rm3_reg_waves "demos/rm3/rm3_reg_waves.cpp")
-add_executable(f3of "demos/f3of/f3of.cpp")
-add_executable(oswec_decay "demos/oswec/oswec_decay.cpp")
+)
+
+add_library(HydroChrono STATIC ${HYDRO_CHRONO_SOURCES} )
+
+
+target_include_directories(
+	HydroChrono
+
+	PUBLIC
+	$<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
+	$<INSTALL_INTERFACE:include>
+
+	PRIVATE
+	${CMAKE_CURRENT_SOURCE_DIR}/
+	${CMAKE_CURRENT_SOURCE_DIR}/include
+
+	)
+
+
+
+
+# add_executable(sphere_reg_waves "demos/sphere/sphere_reg_waves.cpp")
+# add_executable(rm3_decay "demos/rm3/rm3_decay.cpp")
+# add_executable(rm3_reg_waves "demos/rm3/rm3_reg_waves.cpp")
+# add_executable(oswec_decay "demos/oswec/oswec_decay.cpp")
 #SWIG_ADD_LIBRARY(py_wec_chrono LANGUAGE python SOURCES hydro_forces_no_viz.i)
 
 target_compile_features(HydroChrono PUBLIC cxx_std_17)
@@ -84,48 +107,61 @@ set_target_properties(HydroChrono PROPERTIES
 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(sphere_decay PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
+# set_target_properties(sphere_decay PROPERTIES 
+# 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
+# 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
+# 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(sphere_reg_waves PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
+# set_target_properties(sphere_reg_waves PROPERTIES 
+# 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
+# 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
+# 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(rm3_decay PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
+# set_target_properties(rm3_decay PROPERTIES 
+# 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
+# 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
+# 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(rm3_reg_waves PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
+# set_target_properties(rm3_reg_waves PROPERTIES 
+# 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
+# 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
+# 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
-set_target_properties(f3of PROPERTIES 
-		COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
-
-set_target_properties(oswec_decay PROPERTIES 
-	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
-	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
-	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
+# set_target_properties(oswec_decay PROPERTIES 
+# 	    COMPILE_FLAGS "${CHRONO_CXX_FLAGS} ${EXTRA_COMPILE_FLAGS}"
+# 	    COMPILE_DEFINITIONS "CHRONO_DATA_DIR=\"${CHRONO_DATA_DIR}\""
+# 	    LINK_FLAGS "${CHRONO_LINKER_FLAGS}")
 
 #--------------------------------------------------------------
 # Link to Chrono libraries and dependency libraries
 #--------------------------------------------------------------
 
 target_link_libraries(HydroChrono ${LINK_LIBS})
-target_link_libraries(sphere_decay HydroChrono)
-target_link_libraries(sphere_reg_waves HydroChrono)
-target_link_libraries(rm3_decay HydroChrono)
-target_link_libraries(rm3_reg_waves HydroChrono)
-target_link_libraries(f3of HydroChrono)
-target_link_libraries(oswec_decay HydroChrono)
+
+# target_link_libraries(sphere_reg_waves HydroChrono)
+# target_link_libraries(rm3_decay HydroChrono)
+# target_link_libraries(rm3_reg_waves HydroChrono)
+# target_link_libraries(oswec_decay HydroChrono)
 #SWIG_LINK_LIBRARIES(py_wec_chrono wec_chrono ${PYTHON_LIBRARIES})
+
+
+
+
+
+
+
+
+add_executable(sphere_decay demos/sphere/sphere_decay.cpp)
+
+target_link_libraries(sphere_decay HydroChrono)
+
+
+
+
+
+
+
+
 
 #--------------------------------------------------------------
 # === 4 (OPTIONAL) ===
@@ -138,5 +174,5 @@ target_link_libraries(oswec_decay HydroChrono)
 # (this can simply be ${CMAKE_BINARY_DIR}, like in this example)
 #--------------------------------------------------------------
 
-set(EXECUTABLE_OUTPUT_PATH ${CMAKE_BINARY_DIR})
-add_DLL_copy_command("${CHRONO_DLLS}")
+#set(EXECUTABLE_OUTPUT_PATH ${CMAKE_BINARY_DIR})
+#add_DLL_copy_command("${CHRONO_DLLS}")
